// frontend/types/api.ts
export interface components {
    schemas: {
        GetUserResponse: {
            email: string;
            name: string;
            /** Format: double */
            seikin_similarity: number;
            user_id: string;
        };
        RefreshResponse: {
            jwt: string;
            refresh_token: string;
        };
        SignInRequest: {
            email: string;
            password: string;
        };
        SignInResponse: {
            jwt: string;
            refresh_token: string;
        };
        SignUpRequest: {
            email: string;
            name: string;
            password: string;
        };
        SignUpResponse: {
            jwt: string;
            refresh_token: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}